improve docs, export types for extensions, skip integration tests
